# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/materials": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Lấy danh sách tất cả nguyên vật liệu",
                "operationId": "6ec3da580f8b491d0463f51716f29aef",
                "responses": {
                    "200": {
                        "description": "Danh sách nguyên vật liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Nhựa HDPE"
                                            },
                                            "unit": {
                                                "type": "string",
                                                "example": "bao"
                                            },
                                            "quantity": {
                                                "type": "integer",
                                                "example": 50
                                            },
                                            "warehouse_name": {
                                                "type": "string",
                                                "example": "Kho nguyên vật liệu 1"
                                            },
                                            "material_img": {
                                                "type": "string",
                                                "example": null,
                                                "nullable": true
                                            },
                                            "status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-09-25T03:42:10.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": null,
                                                "nullable": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không có nguyên vật liệu nào",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách nguyên vật liệu"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tìm thấy nguyên vật liệu nào"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy danh sách nguyên vật liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy danh sách nguyên vật liệu"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/material/add": {
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "Thêm mới nguyên vật liệu",
                "operationId": "de8f69d24efad037f3e5441166754f50",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "unit",
                                    "quantity",
                                    "category_id",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nhựa HDPE"
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "bao"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "material_img": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thêm nguyên vật liệu thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm nguyên vật liệu thành công"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Thêm nguyên vật liệu thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm nguyên vật liệu thất bại"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/material/{id}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Lấy thông tin nguyên vật liệu",
                "operationId": "c77d3df03ee602c5f1cbdf58737be654",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin nguyên vật liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 3
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Nhựa HDPE"
                                        },
                                        "unit": {
                                            "type": "string",
                                            "example": "bao"
                                        },
                                        "quantity": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "warehouse_name": {
                                            "type": "string",
                                            "example": "Kho nguyên vật liệu 1"
                                        },
                                        "material_img": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-09-25T03:42:10.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy nguyên vật liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin nguyên vật liệu"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Không tim thấy nguyên vật liệu"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi khi lấy thông tin nguyên vật liệu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lỗi khi lấy thông tin nguyên vật liệu"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message here"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}